# apis-service/docs/swagger.yaml
basePath: /
definitions:
...
  types.PlanDto:
    properties:
      apiID:
        type: integer
      description:
        type: string
      features:
        type: string
      limiteAmount:
        type: integer
      limiteTimeUnit:
        type: string
      limiteType:
        type: string
      name:
        type: string
      price:
        type: number
      recomonded:
        type: boolean
      type:
        type: string
      visibility:
        type: string
    type: object
  types.SubscriptionDto:
    properties:
      apiID:
        type: integer
      planId:
        type: integer
    type: object
...
host: localhost:5000
info:
  contact: {}
  description: This is a sample server for a pet store.
  termsOfService: http://terms-of-service-url.com
  title: API Title
  version: "1.0"
paths:
  /Subscription:
    post:
      consumes:
      - application/json
      description: Creates a new API Subscription
      parameters:
      - description: API Subscription Data
        in: body
        name: Subscription
        required: true
        schema:
          $ref: '#/definitions/types.SubscriptionDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create an API Subscription
      tags:
      - API Subscription Operations
  /Subscription/:api-id:
    get:
      description: Retrieves Subscriptions for a specific API
      parameters:
      - description: API ID
        in: path
        name: api-id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get API Subscriptions by API ID
      tags:
      - API Subscription Operations
  /Subscription/:id:
    delete:
      description: Deletes a specific API Subscription
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete an API Subscription
      tags:
      - API Subscription Operations
    patch:
      consumes:
      - application/json
      description: Updates an existing API Subscription
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: integer
      - description: API Subscription Data
        in: body
        name: Subscription
        required: true
        schema:
          $ref: '#/definitions/types.SubscriptionDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update an API Subscription
      tags:
      - API Subscription Operations
...
  /plan:
    post:
      consumes:
      - application/json
      description: Creates a new API Plan
      parameters:
      - description: API Plan Data
        in: body
        name: plan
        required: true
        schema:
          $ref: '#/definitions/types.PlanDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create an API Plan
      tags:
      - API Plan Operations
  /plan/:api-id:
    get:
      description: Retrieves plans for a specific API
      parameters:
      - description: API ID
        in: path
        name: api-id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get API Plans by API ID
      tags:
      - API Plan Operations
  /plan/:id:
    delete:
      description: Deletes a specific API Plan
      parameters:
      - description: Plan ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete an API Plan
      tags:
      - API Plan Operations
    patch:
      consumes:
      - application/json
      description: Updates an existing API Plan
      parameters:
      - description: Plan ID
        in: path
        name: id
        required: true
        type: integer
      - description: API Plan Data
        in: body
        name: plan
        required: true
        schema:
          $ref: '#/definitions/types.PlanDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update an API Plan
      tags:
      - API Plan Operations
schemes:
- http
- https
swagger: "2.0"